Move baaahs.Shader into baaahs.shaders package.
