+++ src/commonMain/kotlin/baaahs/proto/Protocol.kt
@@ -1,11 +1,11 @@
 package baaahs.proto
 
 import baaahs.BrainId
-import baaahs.Shader
 import baaahs.geom.Vector2F
 import baaahs.geom.Vector3F
 import baaahs.io.ByteArrayReader
 import baaahs.io.ByteArrayWriter
+import baaahs.shaders.Shader
 
 object Ports {
     const val PINKY = 8002
